feat: team data submission done style need to be fixed and payment not done

+++ database/migrations/2025_01_19_120104_create_rally_histories_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->uuid('team_id');
             $table->timestamp('qr_expired_at');
             $table->timestamp('scanned_at')->nullable();
-            $table->smallInteger('rank')->nullable()->comment('1: rank 1, 2: rank 2, 3: rank 3');
+            $table->tinyInteger('rank')->nullable()->comment('1: rank 1, 2: rank 2, 3: rank 3');
             $table->unsignedInteger('point')->nullable();
             $table->timestamps();
